requirements.txt
astroid==2.2.5
cycler==0.10.0
decorator==4.4.0
h5py==2.9.0
imageio==2.5.0
isort==4.3.20
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
lxml==4.3.3
matplotlib==3.1.0
mccabe==0.6.1
networkx==2.3
numpy==1.16.4
Pillow==6.0.0
pylint==2.3.1
pyparsing==2.4.0
python-dateutil==2.8.0
PyWavelets==1.0.3
scikit-image==0.15.0
scipy==1.3.0
six==1.12.0
torch==1.1.0
torchvision==0.3.0
typed-ast==1.4.0
wrapt==1.11.1
